remove implicit integer promotion
git-svn-id: https://cieforge.univ-nantes.fr/svn/iatal/trunk@195 399367ef-d520-467f-92a1-83a083ad6548

+++ src/tmx3/parser.cc
@@ -156,7 +156,7 @@ Parser::loadFile(tinyxml2::XMLDocument& tmxDoc, std::string filePath) {
 			std::ostringstream error;
 			error << "Impossible de charger le fichier TMX. "
                   << "Erreur numéro : "
-                  << tmxDoc.ErrorID()
+                  << (int) tmxDoc.ErrorID()
                   << ".";
 			throw XmlException(error.str());
 		}
